NGA-227 database setup - committing for move
