+++ infra_warehouse/postgres/variables.tf
@@ -1,0 +1,12 @@
+
+
+variable "postgres_server_spec"{}
+
+variable "postgres_admin_password"{}
+
+variable "aws_region"{}
+
+variable "public_subnets"{}
+
+variable "public_sg" {}
+
