+++ infra_warehouse/terraform.tfvars
@@ -18,3 +18,6 @@ geoserver_image = "288871573946.dkr.ecr.ap-southeast-2.amazonaws.com/ausseabed-g
 geoserver_initial_memory="2G"
 geoserver_maximum_memory="2G"
 geoserver_admin_password="c1d2-54fe-84f9-7149-f005-ffa8-cbe6-e7b4"
+
+postgres_server_spec="db.t2.micro"
+postgres_admin_password=""
