+++ ingestion/Dockerfile
@@ -3,5 +3,5 @@ ENV APPDIR=/code
 WORKDIR $APPDIR
 
 
-RUN yarn global add mocha typeorm @babel/core @babel/cli @babel/register
+RUN yarn global add mocha typeorm @babel/core reflect-metadata @babel/cli @babel/register pg
