+++ ingestion/package.json
@@ -12,6 +12,8 @@
     "@types/mocha": "^7.0.2",
     "@types/node": "^13.9.2",
     "mocha": "^7.1.1",
+    "pg": "^7.18.2",
+    "reflect-metadata": "^0.1.13",
     "ts-node": "^8.7.0",
     "typeorm": "^0.2.24",
     "typescript": "^3.8.3"
